refactor(admin/languages): move related files into dedicated folder
