+++ src/app/admin/languages/page.tsx
@@ -1,4 +1,4 @@
-import AdminLanguagesPageClient from '@/app/admin/AdminLanguagesPageClient';
+import AdminLanguagesPageClient from '@/app/admin/languages/AdminLanguagesPageClient';
 
 export default function AdminLanguagesPage() {
   return <AdminLanguagesPageClient />;
